Move podDevices code into a separate file.
